requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
gunicorn
